pyproject.toml
[project]
name = "servicehub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.0",
    "jinja2>=3.1.6",
    "pip>=26.0.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
